Refactor auth proxy logic
